+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTarget.java
@@ -960,7 +960,13 @@ public class TextEditingTarget implements
    @Handler
    void onSwitchFocusSourceConsole()
    {
-      if (docDisplay_.isFocused())
+      // Check if either the regular source editor (docDisplay) has focus
+      // or if the visual editor has focus (either ProseMirror or a code chunk)
+      boolean sourceHasFocus = docDisplay_.isFocused() || 
+                              visualEditorHasFocus_ || 
+                              (visualMode_.isActivated() && visualMode_.getActiveEditor() != null);
+      
+      if (sourceHasFocus)
          commands_.activateConsole().execute();
       else
          commands_.activateSource().execute();
@@ -1080,7 +1086,13 @@ public class TextEditingTarget implements
 
    public void onVisualEditorBlur()
    {
+      visualEditorHasFocus_ = false;
       maybeAutoSaveOnBlur();
+   }
+   
+   public void onVisualEditorFocus()
+   {
+      visualEditorHasFocus_ = true;
    }
 
    public void navigateToXRef(String xref)
@@ -9493,6 +9505,7 @@ public class TextEditingTarget implements
    private TextEditingTargetThemeHelper themeHelper_;
    private boolean ignoreDeletes_;
    private boolean forceSaveCommandActive_ = false;
+   private boolean visualEditorHasFocus_ = false;
    private final TextEditingTargetScopeHelper scopeHelper_;
    private final TextEditingTargetCopilotHelper copilotHelper_;
    private TextEditingTargetPackageDependencyHelper packageDependencyHelper_;